pyproject.toml
[tool.poetry]
name = "csv-app"
version = "0.1.0"
description = ""
authors = ["Baumgartner <baumgartner.benedikt@hotmail.com>"]
readme = "README.md"
packages = [{ include = "csv-app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.42.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
openai = "^1.63.2"
asyncio = "^3.4.3"

[tool.poetry.scripts]
csv-app = "streamlit.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"